Eliminate use of unmanaged global variables
Add a new function to the prelude which checks that all globals are
managed, or declared to be managed, in some way. Rewrite all of the
cells to avoid creating unmanaged globals.

I certainly think this makes the notebook cleaner overall, and
encourages creating reusable functions, etc.; however, it does feel a
little like the magical exploratory nature of the notebook medium has
been deadened a bit. We'll see if I keep using this.

If nothing else, I can always use the `assert_globals_clean` function
just for *checking* the globals so that I can delete garbage once in a
while.